Specified the type for the patch offset array stored by the patch set.
The array defaulst to a floating point type, which is incorrect, and
causes problems when attempting to use this for element access.

+++ modules/sample.py
@@ -243,7 +243,7 @@ class PatchSet( object ):
         patch_offsets = numpy.array(
             [ ( v + volume_offset, z, y, x )
               for v, offsets in enumerate( patch_offsets_per_volume )
-              for z, y, x in offsets ] )
+              for z, y, x in offsets ] ).astype( 'int64' )
         self.__patch_offsets = patch_offsets
 
         log.item( "extracting image patches" )
